Fixed filename bug
I tested this code and noticed that every time I save a file, no matter
which feature (body part) I chose, the feature portion of the filename
of the csv file is always "DOMINANT_WRIST". So, I've fixed it. I tested
this and now filename contains the feature you chose on the watch.